Dynamically load file from source to sync
