+++ python-sdk/src/transfers/data_providers/aws/s3.py
@@ -9,9 +9,10 @@ from transfers.constants import FileLocation, TransferMode
 from transfers.data_providers import create_dataprovider
 from transfers.data_providers.base import DataProviders
 from transfers.datasets.base import UniversalDataset as Dataset
-from transfers.utils import get_dataset_connection_type
 
 from astro.constants import LoadExistStrategy
+
+# from transfers.utils import get_dataset_connection_type
 
 
 class S3DataProvider(DataProviders):
@@ -35,10 +36,12 @@ class S3DataProvider(DataProviders):
             if_exists=if_exists,
         )
         self.transfer_mapping: set = {
-            FileLocation.GS,
-            FileLocation.S3,
             FileLocation.AWS,
             FileLocation.google_cloud_platform,
+        }
+        self.LOAD_DATA_FROM_SOURCE = {
+            "google_cloud_platform": "load_data_from_gcs",
+            "aws": "load_data_from_s3",
         }
 
     @property
@@ -54,23 +57,6 @@ class S3DataProvider(DataProviders):
     def check_if_exists(self, dataset: Dataset) -> bool:
         """Return true if the dataset exists"""
         raise NotImplementedError
-
-    def check_if_transfer_supported(self, source_dataset: Dataset) -> bool:
-        """
-        Checks if the transfer is supported from source to destination based on source_dataset.
-        """
-        source_connection_type = get_dataset_connection_type(source_dataset)
-        return FileLocation(source_connection_type) in self.transfer_mapping
-
-    def load_data_from_source(self, source_dataset: Dataset, destination_dataset: Dataset) -> None:
-        """
-        Loads data from source dataset to the destination using data provider
-        """
-        if not self.check_if_transfer_supported(source_dataset=source_dataset):
-            raise ValueError("Transfer not supported yet.")
-        source_connection_type = get_dataset_connection_type(source_dataset)
-        if source_connection_type == "google_cloud_platform":
-            return self.load_data_from_gcs(source_dataset, destination_dataset)
 
     def load_data_from_gcs(self, source_dataset: Dataset, destination_dataset: Dataset) -> None:
         source_dataprovider = create_dataprovider(
@@ -129,6 +115,9 @@ class S3DataProvider(DataProviders):
             logging.info("In sync, no files needed to be uploaded to S3")
 
         return files
+
+    def load_data_from_s3(self, source_dataset: Dataset, destination_dataset: Dataset) -> None:
+        raise NotImplementedError
 
     def get_bucket_name(self, dataset: Dataset) -> str:
         bucket_name, key = self.hook.parse_s3_url(dataset.path)
