+++ python-sdk/src/transfers/data_providers/google/cloud/gcs.py
@@ -5,7 +5,6 @@ from airflow.providers.google.cloud.hooks.gcs import GCSHook, _parse_gcs_url
 from transfers.constants import FileLocation, TransferMode
 from transfers.data_providers.base import DataProviders
 from transfers.datasets.base import UniversalDataset as Dataset
-from transfers.utils import get_dataset_connection_type
 
 from astro.constants import LoadExistStrategy
 
@@ -31,10 +30,12 @@ class GCSDataProvider(DataProviders):
             if_exists=if_exists,
         )
         self.transfer_mapping: set = {
-            FileLocation.GS,
-            FileLocation.S3,
             FileLocation.AWS,
             FileLocation.google_cloud_platform,
+        }
+        self.LOAD_DATA_FROM_SOURCE = {
+            "google_cloud_platform": "load_data_from_gcs",
+            "aws": "load_data_from_s3",
         }
 
     @property
@@ -54,24 +55,10 @@ class GCSDataProvider(DataProviders):
         bucket_name, blob = _parse_gcs_url(gsurl=source_dataset.path)
         return bucket_name
 
-    def load_data_from_source(self, source_dataset: Dataset, destination_dataset: Dataset) -> None:
-        """
-        Loads data from source dataset to the destination using data provider
-        """
-        if not self.check_if_transfer_supported(source_dataset=source_dataset):
-            raise ValueError("Transfer not supported yet.")
-        source_connection_type = get_dataset_connection_type(source_dataset)
-        if source_connection_type == "google_cloud_platform":
-            return self.load_data_from_gcs(source_dataset, destination_dataset)
-
-    def check_if_transfer_supported(self, source_dataset: Dataset) -> bool:
-        """
-        Checks if the transfer is supported from source to destination based on source_dataset.
-        """
-        source_connection_type = get_dataset_connection_type(source_dataset)
-        return source_connection_type in self.transfer_mapping
-
     def load_data_from_gcs(self, source_dataset: Dataset, destination_dataset: Dataset) -> None:
+        raise NotImplementedError
+
+    def load_data_from_s3(self, source_dataset: Dataset, destination_dataset: Dataset) -> None:
         raise NotImplementedError
 
     @property
